changed version to __version__, added checklist
I changed the variable version to __version__ so Kivy knows what the version is when compiling. It also feels better, probably because it's kind of a special variable.
I added a checklist in doc/ so that we don't forget anything when releasing a new version. Feel free to add anything here.