updated licence and logo printing mechanism
I experimented a little bit to find out the best printing mechanism
for the logo and license. Both of these are text-heavy, and the blocks of
text decreased the readability of kotnetcli.py.

One approach is to put the information in logo.txt, and then read it
out line by line. This works, but we want some text to be variable, being
the Github URL and the version number. Hardcoding this information is not
feasible, there are lots of files that contain the version number and we
can't possibly replace all of them every time we bump up the release number.
So it should be dynamic instead.

I was thinking about replacing the contents in logo.txt. AFAIK, the only way
to do this is by replacing by regex or by trying to format every line, also
the lines which don't contain a to-be-replaced string. That will probably work
but is a messy approach in my opinion, I don't want kotnetcli to try to
find a {version} string in a line when it's not there.

The solution I went with is to create two triple-quoted strings, which
can be formatted really easily. It's very clear to Python what should be
formatted, because every line is in one big string, so the string contains
everything the format() keyword needs.

The strings reside in logo.py and license.py inside the tools/ directory.
We can probably move them later on in one file or create a separate directory
for this.