Explicit communicator methods for each kind of failure.
This refactors the single parameterized method into a set of _explicit_
methods for each kind of expected failure, with additional info passed
as parameters to the communicator. Likewise, _unexpected_ failure (i.e.
exceptions) are handled via a single 'eventFailureInternalError'
communicator method that takes the traceback as an argument.

This nicely decouples visualisation concerns from worker functionality,
and ensures all kinds of failures are handled gracefully.