Catch communicatorFabriek lazy importing exceptions in front-end.
The lazy importing mechanism only imports a communicator when it has been
requested by the front-end, via the fabriek. This ensures an end-user
does not have to install all the dependencies for all the communicators
in order to run kotnetcli. The front-end is thus responsible to catch
any ImportErrors thrown at communicator creation time.

This nicely maintains the behavior that *only* the front-end and the
communicators encapsulate user interaction + only the front-end or
worker may terminate (sys.exit) the program.

This fixes issue #9 .