Re-implementation of the Dialog communicator.
The dialog communicator draws a screen with an overview of the
different steps of the netlogin process. The implementation keeps
track of the current step (to be updated on the next event) via a
custom login/logout/forcelogin-specific iterator object.

Having the dialog dependency lazy-imported via the Fabriek, allows
kotnetcli to be used without having to install dialog (issue #43 ).