Connect to localhost development server over plain HTTP.
An SSL-enabled connection for a localhost test server is pretty
useless anyway, and the exception for the self-signed certificate
polluted the code base.

Note that the browser class of course uses a properly verified
HTTPS connection for the legitimate netlogin server. We can always
re-enable SSL in our test server, should we ever want to deploy a
remote netlogin test service.